BINIUS-234: replace the smul gate with a standard gadget
Signed 64-bit multiplication was a bespoke opcode carrying seven aux
wires, a scratch register, its own hand-rolled modular-addition
constraint helper, a dedicated bytecode instruction, and a matching
interpreter path. All of that existed only to express something the
primitive gates already can.

Introduce a `binius_frontend::ops` module holding an `smul64` gadget
that composes existing gates: one unsigned multiply for the 128-bit
product, two arithmetic shifts for the sign masks, two ANDs for the
sign corrections, and two subtractions for the high-word fix-up. The
low word is the unsigned low word unchanged. `CircuitBuilder::smul` is
reopened in the same module as a thin wrapper over the gadget.

The `Smul` opcode and gate, its `emit_smul` bytecode, and the
`exec_smul` interpreter path are deleted. The full test suite moves to
the gadget module.
