+++ crates/frontend/src/compiler/gate/mod.rs
@@ -36,7 +36,6 @@ pub mod select;
 pub mod shl;
 pub mod shr;
 pub mod sll32;
-pub mod smul;
 pub mod sra32;
 pub mod srl32;
 
@@ -66,7 +65,6 @@ pub fn constrain(gate: Gate, graph: &GateGraph, builder: &mut ConstraintBuilder)
 		Opcode::AssertTrue => assert_true::constrain(gate, data, builder),
 		Opcode::AssertEqCond => assert_eq_cond::constrain(gate, data, builder),
 		Opcode::Imul => imul::constrain(gate, data, builder),
-		Opcode::Smul => smul::constrain(gate, data, builder),
 		Opcode::IcmpUlt => icmp_ult::constrain(gate, data, builder),
 		Opcode::IcmpEq => icmp_eq::constrain(gate, data, builder),
 		Opcode::Shr => shr::constrain(gate, data, builder),
@@ -130,7 +128,6 @@ pub fn emit_gate_bytecode(
 			assert_true::emit_eval_bytecode(gate, data, assertion_path, builder, wire_to_reg)
 		}
 		Opcode::Imul => imul::emit_eval_bytecode(gate, data, builder, wire_to_reg),
-		Opcode::Smul => smul::emit_eval_bytecode(gate, data, builder, wire_to_reg),
 		Opcode::IcmpUlt => icmp_ult::emit_eval_bytecode(gate, data, builder, wire_to_reg),
 		Opcode::IcmpEq => icmp_eq::emit_eval_bytecode(gate, data, builder, wire_to_reg),
 		Opcode::Shr => shr::emit_eval_bytecode(gate, data, builder, wire_to_reg),
